requirements.txt
Flask
openai
python-dotenv
PyMuPDF
requests
beautifulsoup4
chromadb
langchain
docx
